+++ Release_Notes.md
@@ -1,4 +1,7 @@
-﻿# 2025-12-21 - V3 XLSB-Beta
+﻿# 2025-12-31 - V3 XLSB-Beta
+- ✅ Read `definedName`s (Ranges / Cell / Value / Dynamic) 📇
+
+# 2025-12-21 - V3 XLSB-Beta
 - ⛓️‍💥 **Breaking Change(s)
     - `FileType` has been removed, and Open via the Public class type
     - `IXmlReaderHelpers` has become `IOpenXmlReaderHelpers`, with slightly different methods
@@ -10,10 +13,10 @@
     - ✅ Delayed read - until a cell is actually needed
     - ✅ Deal with Null / Empty cells
 - ✅ Cell object type 📅
-    - [>] 👟 [2nd Pass Performance on 2025-12-21](https://github.com/Smurf-IV/Excel_PRIME/blob/main/Performance.md#2025-12-21)
+    - ✅ 👟 [2nd Pass Performance on 2025-12-21](https://github.com/Smurf-IV/Excel_PRIME/blob/main/Performance.md#2025-12-21)
 
 
-# 2025-12-10 - RC
+# 2025-12-10 - V2 RC
 - User defined, using the `"A1:B10"` or `"$A$1:$B$10"` syntax
   - [Range Performance on 2025-12-10](https://github.com/Smurf-IV/Excel_PRIME/blob/main/Performance.md#2025-12-10)